Modified templates to wrap lines at 72 characters
Some popular mail client PGP software will wrap lines at 72
characters.  If we send an OSSN out that has longer lines,
the formatting gets messed up.  This can make the OSSN hard
to read.  This patch modifies the templates to wrap lines at
72 characters so they match the guidelines posted in the OSSN
process pages on the wiki.

Now that we are publishing OSSNs on the wiki, I changed the
templates to include a link to the OSSN on the wiki instead
of duplicating the Launchpad link like we did previously.