Remove wiki template
The wiki template has confused authors with regards to which
template they should use to author a new note.  Since the source
repository only contains notes in the "email" template format, we
should only contain that template here.  The wiki conversion can
be described on the OSSN process page instead.

Change-Id: I4a13ef98bcbcc8a49f14bf30096fb19b0afc8082